+++ flying-saucer-swing/src/main/java/com/github/danfickle/flyingsaucer/swing/test/AllPageTest.java
@@ -1,6 +1,6 @@
-package org.xhtmlrenderer.test;
+package com.github.danfickle.flyingsaucer.swing.test;
 
-import org.xhtmlrenderer.simple.Graphics2DRenderer;
+import com.github.danfickle.flyingsaucer.swing.Graphics2DRenderer;
 import org.xhtmlrenderer.util.Uu;
 
 import java.io.File;
