+++ flying-saucer-swing/src/main/java/com/github/danfickle/flyingsaucer/swing/test/DocumentDiffGenerate.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  * }}}
  */
-package org.xhtmlrenderer.test;
+package com.github.danfickle.flyingsaucer.swing.test;
 
 import java.io.File;
