+++ flying-saucer-swing/src/main/java/com/github/danfickle/flyingsaucer/swing/test/SwingImageReplacer.java
@@ -1,4 +1,4 @@
-package org.xhtmlrenderer.test;
+package com.github.danfickle.flyingsaucer.swing.test;
 
 import org.jsoup.nodes.Element;
 import org.xhtmlrenderer.extend.ReplacedElement;
